* Switches from int to decimal * Extracts TokenFactory class and tests * Extracts tests per Token type

+++ ShuntingYard.Core/Tokens/Operand.cs
@@ -4,5 +4,5 @@ public class Operand : Token
 {
     public Operand(string value) => Value = value;
 
-    public int ToInt() => int.Parse(Value);
+    public decimal ToDecimal() => decimal.Parse(Value);
 }
